libxl: react correctly to POLLHUP
When received POLLHUP on datacopier_readable/writable, kill the
datacopier and call the callback with onwrite=-2. On the bootloader
callback kill the bootloader process and wait for the callback, but
without setting the bl->rc error code.

This is because NetBSD returns POLLHUP when the process on the other
end of the pty exits (even when exiting normally).

Cc: Ian Jackson <ian.jackson@eu.citrix.com>