libxl: convert libxl__device_disk_local_attach to an async op
This will be needed in future patches, when libxl__device_disk_add
becomes async also. Create a new status structure that defines the
local attach of a disk device and use it in
libxl__device_disk_local_attach.

This is done in this patch to split the changes introduced when
libxl__device_disk_add becomes async.

Changes since v9:

 * Fixed state changes comments on header of functions.

 * Rework libxl__device_disk_local_init to just set rc to 0, and
   perform the initialization in the callers context.

 * If attach fails, perform a detach, so the state of the local_device
   on the callback is Attached if rc == 0, or Idle if rc != 0.

 * Free diskpath inside the detach callback.

Changes since v6:

 * Added better comments.

 * Instead of passing the xs_transaction arround several functions,
   put all the xs related operations inside device_disk_add.

 * Make sure libxl__device_disk_local_initiate_detach either calls the
   async device destroy function, or the callback.

Cc: Ian Jackson <ian.jackson@eu.citrix.com>
Cc: Stefano Stabellini <stefano.stabellini@eu.citrix.com>