libxl: add support for different hotplug interfaces
Add a new variable to libxl_device_disk "hotplug_version", that will
be automatically set to the detected hotplug script interface version.

This patch adds the basic support to handle this new scripts, by
adding the prepare/unprepare private libxl functions, and modifying
the current domain creation/destruction to call this functions at the
appropriate spots. The hotplug interface detection is done in the
_prepare call.

This patch also adds some helpers, that will be used here and in later
patches.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Cc: Ian Campbell <ian.campbell@citrix.com>