libxl: add Qdisk backend launch helper
Current Qemu launch functions in libxl require the usage of data 
structures only avaialbe on domain creation. All this information is 
not need in order to launch a Qemu instance to serve Qdisk backends, 
so introduce a new simplified helper that can be used to launch 
Qemu/Qdisk, that will be used to launch Qdisk in driver domains.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Cc: Ian Jackson <Ian.Jackson@eu.citrix.com>
Cc: Ian Campbell <ian.campbell@citrix.com>
Cc: Anthony PERARD <anthony.perard@citrix.com>
Cc: Stefano Stabellini <stefano.stabellini@eu.citrix.com>