libxl: synchronize device removal when using driver domains
Synchronize the clean up of the backend from the toolstack domain when
the driver domain has actually finished closing the backend for the
device.

This is accomplished by waiting for the driver domain to  remove the
directory containing the backend keys, then the toolstack domain will
finish the cleanup by removing the empty folders on the backend path.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>
Cc: Ian Jackson <Ian.Jackson@eu.citrix.com>
Cc: Ian Campbell <ian.campbell@citrix.com>
---
Changes since v1:
 * Place cleanup of aodev events in device_hotplug_clean, get rid of
   some code duplication.

Changes since RFC:
 * This patch has been reworked to synchronize the toolstack and the
   driver domain, by making the driver domain only remove the first
   directory of the backend path, and the toolstack domain remove the
   rest.