x86/hap: use the right cache attributes when MTRR is disabled
Currently the code that calculates the cache attributes of the HAP page
tables assume that if MTRR are disabled the memory type is UC, this can
cause issues if MTRR are never enabled because the guest only plans to use
PAT.

In order to solve this modify epte_get_entry_emt so that is takes into
account that MTRR can be disabled and that PAT should be used instead, this
also implies that when page tables are enabled inside the guest a
recalculation of the HAP page table cache attributes must be performed.

The special casing that was previously done for PVHv1 guests can now be
removed, since PVHv1 guest cannot have MTRR enabled or paging disabled, so
the memory type is always going to be MTRR_TYPE_WRBACK in that case.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
---
Changes since v1:
 - Fix comments.
 - Slightly change the if ... else logic.
 - Correctly check if MTRR are enabled.
 - Expand description.